requirements.txt
gymnasium>=0.29.1
gym>=0.26.2
mujoco>=3.1.5
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
opencv-python>=4.8.0
tqdm>=4.66.0
huggingface_hub>=0.24.0
stable-baselines3>=2.3.0
sb3-contrib>=2.3.0
imageio>=2.37.0
PyYAML>=6.0
Pillow>=10.0.0
shimmy>=0.2.1
